Fixed Peek/GoToDefinition
A recent VSCode update apparently changed how PeekDefinition/GoToDefinition work, which has already negatively impacted ORAIDE. It looks like with that update VSCode started opening and closing files where a definition is when you try to "peek" it. The result here is that `onDidCloseTextDocument` is called for the peeked file when it shouldn't and the generated map file (which is the one being peeked) gets erroneously deleted.
Adding a check whether the original file (mission script) or the peeked file (generated based on map.yaml) is being closed fixes that.